Use a random port for the test etcd server

+++ hack/lib/etcd.sh
@@ -16,20 +16,21 @@
 
 # A set of helpers for starting/running etcd for tests
 
+# Sets ETCD_HOST and ETCD_PORT if not already set.
+# Also sets ETCD_PID and ETCD_DIR and installs an EXIT trap for cleanup.
 kube::etcd::start() {
-  local host=${ETCD_HOST:-127.0.0.1}
-  local port=${ETCD_PORT:-4001}
+  ETCD_HOST=${ETCD_HOST:-127.0.0.1}
+  ETCD_PORT=${ETCD_PORT:-$(kube::util::get_random_port)}
+
+  if [[ -n "${ETCD_PID-}" ]]; then
+    kube::log::error "etcd already started with pid ${ETCD_PID}!"
+    exit 1
+  fi
 
   which etcd >/dev/null || {
     kube::log::usage "etcd must be in your PATH"
     exit 1
   }
-
-  if pgrep etcd >/dev/null 2>&1; then
-    kube::log::usage "etcd appears to already be running on this machine (`pgrep -l etcd`) (or its a zombie and you need to kill its parent)."
-    kube::log::usage "retry after you resolve this etcd error."
-    exit 1
-  fi
 
   version=$(etcd -version | cut -d " " -f 3)
   if [[ "${version}" < "2.0.0" ]]; then
@@ -39,22 +40,29 @@ kube::etcd::start() {
 
   # Start etcd
   ETCD_DIR=$(mktemp -d 2>/dev/null || mktemp -d -t test-etcd.XXXXXX)
-  kube::log::info "etcd -data-dir ${ETCD_DIR} --bind-addr ${host}:${port} >/dev/null 2>/dev/null"
-  etcd -data-dir ${ETCD_DIR} --bind-addr ${host}:${port} >/dev/null 2>/dev/null &
+  kube::log::info "Starting etcd with data-dir ${ETCD_DIR} and bind-addr ${ETCD_HOST}:${ETCD_PORT}"
+  etcd -data-dir ${ETCD_DIR} --bind-addr ${ETCD_HOST}:${ETCD_PORT} >/dev/null 2>/dev/null &
   ETCD_PID=$!
+  kube::util::trap_add kube::etcd::cleanup EXIT
 
   echo "Waiting for etcd to come up."
-  kube::util::wait_for_url "http://${host}:${port}/v2/machines" "etcd: " 0.25 80
-  curl -fs -X PUT "http://${host}:${port}/v2/keys/_test"
+  kube::util::wait_for_url "http://${ETCD_HOST}:${ETCD_PORT}/v2/machines" "etcd: " 0.25 80
+  curl -fs -X PUT "http://${ETCD_HOST}:${ETCD_PORT}/v2/keys/_test"
 }
 
 kube::etcd::stop() {
-  kill "${ETCD_PID-}" >/dev/null 2>&1 || :
-  wait "${ETCD_PID-}" >/dev/null 2>&1 || :
+  if [[ -n "${ETCD_PID-}" ]]; then
+    kill "${ETCD_PID-}" >/dev/null 2>&1 || :
+    wait "${ETCD_PID-}" >/dev/null 2>&1 || :
+    ETCD_PID=
+  fi
 }
 
 kube::etcd::clean_etcd_dir() {
-  rm -rf "${ETCD_DIR-}"
+  if [[ -n "${ETCD_DIR-}" ]]; then
+    rm -rf "${ETCD_DIR-}"
+    ETCD_DIR=
+  fi
 }
 
 kube::etcd::cleanup() {
